requirements.txt
openai~=0.27.8
fastapi~=0.100.1
python-dotenv~=1.0.0
pydantic~=2.1.1
locust~=2.15.1